issue/#16

+++ backend/database/migrations/2023_12_05_150912_create_posts_table.php
@@ -11,14 +11,16 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('posts', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('user_id')->nullable();
-            $table->text('text')->nullable();
-            $table->string('image')->nullable();
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('set null');
-            $table->timestamps();
-        });
+        if(!Schema::hasTable('posts')){ 
+            Schema::create('posts', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('user_id')->nullable();
+                $table->text('text')->nullable();
+                $table->string('image')->nullable();
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('set null');
+                $table->timestamps();
+            });
+        }  
     }
 
     /**
